Explicitly casting a narrowing conversion

+++ source/1.Getting_Started/3.Shaders/2.Green_To_Black_Ex2.cpp
@@ -134,7 +134,7 @@ int main() {
         glClearColor(0.102f, 0.110f, 0.118f, 1.0f);
         glClear(GL_COLOR_BUFFER_BIT);
 
-        float timeValue = glfwGetTime();
+        float timeValue = (float)glfwGetTime();
         float greenValue = (sin(timeValue) / 2.0f) + 0.5f;
         int vertexColorLocation =
                 glGetUniformLocation(shaderProgram, "ourColor");
